+++ src/components/Todo/StatusToggler.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+type Props = {
+  completed: boolean;
+  onToggle?: () => void;
+};
+
+export const StatusToggler: React.FC<Props> = ({
+  completed,
+  onToggle = () => { },
+}) => {
+  return (
+    <label className="todo__status-label">
+      <input
+        type="checkbox"
+        className="todo__status"
+        onChange={onToggle}
+        checked={completed}
+      />
+    </label>
+  );
+};
