+++ src/components/common/index.ts
@@ -1,8 +1,5 @@
 import { Notification } from './Notification';
 import { Form } from './Form';
-import { TodoCard } from './TodoCard';
 import { Toggler } from './Toggler';
 
-export {
-  Notification, Form, TodoCard, Toggler,
-};
+export { Notification, Form, Toggler };
